added testing tweaks for more code coverage
