+++ tests/storage/annotations/Swagger/Anotations.php
@@ -16,6 +16,15 @@
  * )
  */
 
+/**
+ * @SWG\SecurityScheme(
+ *   securityDefinition="passport",
+ *   type="oauth2",
+ *   tokenUrl="/oauth/token",
+ *   flow="password",
+ *   scopes={}
+ * )
+ */
 
 /**
  * @SWG\Get(
